feat: add useElementInViewportProgress hook
